Redirect to Location with a URL hash
This pull request adds test coverage to demonstrate a limitation in
Turbo's support for redirecting to a URL with a [hash][] value.

Currently, redirecting to a URL with a hash (for example,
`/src/tests/fixtures/one.html#element-id`) omits the hash
(`#element-id`) when navigating to the final destination.

According to the [fetch specification][], redirects should preserve URL
fragments during redirection.

[hash]: https://developer.mozilla.org/en-US/docs/Web/API/URL/hash
[fetch specification]: https://fetch.spec.whatwg.org/#http-redirect-fetch